update scripts, add some path verification

+++ scripts/backup.sh
@@ -1,7 +1,14 @@
+PROJECT_DIR="~/TokenDroplet"
+
 for i in "$@"; do
 
-PROJECT_DIR="~/TokenDroplet"
 echo "backup $i"
+if [ ! -e "${PROJECT_DIR}/$i/ScanService/config/service.json" ]; then
+    echo "${PROJECT_DIR}/$i/ScanService/config/service.json does not exist!"
+    continue
+fi
+
+mkdir -p ${PROJECT_DIR}/config/ScanService
 cp -v ${PROJECT_DIR}/$i/ScanService/config/service.json ${PROJECT_DIR}/config/ScanService/$i.json
 
 done
